Paging and cursor-based search and retrieval has been added and i have added a column in the event entity about the attendee count

+++ src/main/java/jpolanco/springbootapp/event/application/services/DeleteMyEventById.java
@@ -4,6 +4,7 @@ import jpolanco.springbootapp.event.application.errors.EventAppError;
 import jpolanco.springbootapp.event.application.errors.EventPersistenceFailure;
 import jpolanco.springbootapp.event.application.ports.output.EventRepository;
 import jpolanco.springbootapp.event.application.uc.DeleteMyEventByIdUC;
+import jpolanco.springbootapp.event.domain.model.EventStatus;
 import jpolanco.springbootapp.event.infrastructure.services.interfaces.ImageStorageService;
 import jpolanco.springbootapp.shared.domain.Result;
 import lombok.RequiredArgsConstructor;
@@ -23,9 +24,13 @@ public class DeleteMyEventById implements DeleteMyEventByIdUC {
         if (!event.get().getCreatorId().equals(creatorId)) {
             return Result.failure(EventAppError.EVENT_NOT_BELONG_TO_USER);
         }
+        // Check if the event has already been cancelled or completed
+        if (!(event.get().getStatus().equals(EventStatus.CANCELLED) || event.get().getStatus().equals(EventStatus.COMPLETED))) {
+            return Result.failure(EventAppError.EVENT_NOT_CANCELLED_OR_COMPLETED);
+        }
         eventRepository.deleteByIdAndCreatorId(eventId, creatorId);
         if (!imageStorageService.deleteImage(event.get().getPictureFileName()))
-            throw new EventPersistenceFailure("Failed to delete event image: " + event.get().getPictureFileName());
+            return Result.failure(EventAppError.IMAGE_DELETE_ERROR);
         return Result.success();
     }
 }
